added better signalhandling when interactive
